+++ rtl/mkidaranth/src/mkidaranth/image_cuber_perhipheral.py
@@ -6,7 +6,7 @@ from amaranth.utils import exact_log2
 from amaranth_soc import csr
 
 from .trigger import trigger_event, CYCLE_BITS
-from .image_cuber import ImageCuber, Producer
+from .image_cuber import ImageCuber
 from . import axi
 
 class AddressGenerator(wiring.Component):
@@ -227,53 +227,4 @@ class CuberPeri(wiring.Component):
             m.d.sync += self.membus.ar.ready.eq(1)
 
         return m
-    
 
-
-class Harness(Component):
-    test_phase: In(signed(16))
-    test_bin: In(11)
-    test_cycle: In(CYCLE_BITS)
-    photon_event: In(1)
-    valid: In(1)
-    membus: In(axi.Signature(axi.Axi4Properties(QOS_Present=False, PROT_Present=False, CACHE_Present=False, Exclusive_Accesses=False, READ_WRITE_MODE=axi.ReadWriteMode.READ_ONLY, ADDR_WIDTH=16, REGION_Present=False, DATA_WIDTH=64, WSTRB_Present=False, WLAST_Present=False, ID_W_WIDTH=0, ID_R_WIDTH=2)))
-    cycle_counter: Out(CYCLE_BITS)
-    
-    def __init__(self):
-        super().__init__()
-
-    def elaborate(self, platform):
-        m = Module()
-
-        m.d.sync += self.cycle_counter.eq(self.cycle_counter + 1)
-
-        m.submodules.producer = producer = Producer()
-        m.submodules.cuber_peri = cuber_peri = CuberPeri(addr_width=8, data_width=16)
-
-        self.cuber_peri = cuber_peri
-
-        m.d.comb += [
-            cuber_peri.trigger_stream.payload.eq(producer.event_stream.payload),
-            cuber_peri.trigger_stream.valid.eq(producer.event_stream.valid),
-            producer.event_stream.ready.eq(cuber_peri.trigger_stream.ready)
-        ]
-
-        wiring.connect(m, wiring.flipped(self.membus), cuber_peri.membus)
-
-        cycle_last = Signal(CYCLE_BITS)
-
-        with m.If(self.photon_event):
-            m.d.comb += producer.event_stream.payload.cycle.eq(self.test_cycle)
-            m.d.sync += cycle_last.eq(self.test_cycle)
-            m.d.sync += self.photon_event.eq(0)
-        with m.Else():
-            m.d.comb += producer.event_stream.payload.cycle.eq(cycle_last)
-
-        m.d.comb += producer.event_stream.payload.phase.eq(self.test_phase)
-        m.d.comb += producer.event_stream.payload.bin.eq(self.test_bin)
-        m.d.comb += producer.event_stream.valid.eq(self.valid)
-        
-        with m.If(producer.event_stream.ready == 1):
-            m.d.sync += self.valid.eq(0)
-
-        return m
